requirements.txt
asgiref==3.5.2
astroid==2.11.7
autopep8==1.7.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
cryptography==37.0.4
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==1.0.0
dj3-cloudinary-storage==0.0.6
Django==3.2.15
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-phonenumber-field==6.3.0
django-summernote==0.8.20.0
Flask==2.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
phonenumberslite==8.12.54
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pylint==2.14.5
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
urllib3==1.26.11
webencodings==0.5.1
Werkzeug==2.2.1
wrapt==1.14.1
whitenoise===6.2.0
django-bootstrap-datepicker-plus==4.0.0